fix: optimise throttle and hooks
In the process of fixing some hook dependencies, I introduced a bug
which caused the application to rapidly send requests to the Krist
server (luckily I was using a local Krist instance). To raise awareness
of this if it happens again in the future, I added a notification that
appears whenever the rate limit is hit. Ideally, it should never appear
in production.